requirements.txt
# Основные зависимости FastAPI
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose>=3.3.0
passlib>=1.7.4

# База данных
SQLAlchemy==2.0.27
alembic==1.13.1

# WebSocket и HTTP клиент
websockets==12.0
aiohttp==3.9.3
backoff==2.2.1

# Планировщик задач
APScheduler>=3.8.0

# Утилиты
python-dateutil==2.8.2
pydantic>=1.8.0
numpy>=1.24.0

# Логирование и мониторинг
structlog==24.1.0
logging==0.4.9.6

# Тестирование
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
pytest-cov==4.1.0

# Линтеры и форматтеры
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# Дополнительные зависимости
bcrypt>=3.2.0
aiosqlite==0.19.0
python-dotenv==1.0.1 